requirements.txt
tensorflow==2.9.3
uvicorn
fastapi
tqdm==4.64.1
Pillow
numpy==1.23.4
pandas==1.4.4
opencv-python==4.7.0.72
scikit-learn==1.2.2
google==2.0.3
streamlit==1.11.1
python-multipart
